this shit should be double for the best accurate results if we are using pi

+++ lucid/lucid/renderer/renderer.cpp
@@ -422,17 +422,17 @@ void lucid_engine::renderer::gradient_triangle(const vec2_t pos, const vec2_t si
 std::vector<vec2_t> lucid_engine::renderer::generate_circle_points(const vec2_t pos, const int radius, const int completion, const int rotation, int segments) {
 	std::vector<vec2_t> points;
 
-	float ang = rotation * D3DX_PI / 180;
-	float c = (completion / 100.f) * D3DX_PI;
+	double ang = static_cast<double>( rotation * D3DX_PI ) / 180.0;
+	double c = (completion / 100.0) * D3DX_PI;
 
 	if (segments == -1)
 		segments = std::clamp(radius, 8, 128);
 
 	for (int i = 0; i < segments + 1; i++) {
-		float current_point = static_cast<float>(i) / segments;
+		double current_point = static_cast<double>(i / segments);
 
-		float x = pos.x + radius * cos(ang + 2 * c * current_point);
-		float y = pos.y + radius * sin(ang + 2 * c * current_point);
+		double x = pos.x + radius * cos(ang + 2 * c * current_point);
+		double y = pos.y + radius * sin(ang + 2 * c * current_point);
 
 		points.emplace_back(vec2_t(x, y));
 	}
